More database schema, initial split of routes into separate files
